requirements.txt
openai>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
qdrant-client>=1.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
httpx>=0.24.0